ce-command-num-print: create
this adds the n command to print with line numbers.

line numbers allow for better ease of use, however this is slower than
normal printing.

the line numbers pad to 3 digits, as extremely large files are a bit out
of scope with commoned (especially with everything being a linked list)